Split utility group into files
