+++ src/impl/utils.ts
@@ -2,7 +2,7 @@ import { unreachableCase } from "ts-assert-unreachable";
 import { Connection, Parameters } from "../sqlite3";
 import { TableColumn, TableDefinition, ForeignKey } from "../undo-log-tables";
 import { UndoLogError } from "../undo-log";
-import { ColumnValue, Utils } from "../utils";
+import { ColumnValue, Utils } from "../utilities";
 
 export interface PragmaTableInfo {
   cid: number;
