+++ src/index.ts
@@ -5,7 +5,8 @@ import { Connection } from "./sqlite3";
 import { Delta, UndoLog } from "./undo-log";
 import { UndoLogAssertions } from "./undo-log-assertions";
 import { UndoLogSetup } from "./undo-log-setup";
-import { createUndoLogUtilityServices, UndoLogUtilityServices } from "./undo-log-utils";
+import { createUndoLogUtilityServices } from "./utils/module";
+import { UndoLogUtilityServices } from "./utils/undo-log-utility-services";
 import { AssertionsImpl } from "./impl/assertions";
 import { UndoLogImpl } from "./impl/undo-log";
 import { UndoLogAssertionsImpl } from "./impl/undo-log-assertions";
