--- src/tests/index.test.ts
@@ -1,26 +1,0 @@
-import { UndoLogPublic, UndoLogSetupPublic } from "..";
-import { Connection } from "../sqlite3";
-import { UndoLogAssertions } from "../undo-log-assertions";
-import { UndoLogUtilityServices } from "../undo-log-utils";
-import { setupBeforeEach } from "./fixtures";
-
-describe("Public API single channel", () => {
-    let connection: Connection;
-    let logUtils: UndoLogUtilityServices;
-    let logAssertions: UndoLogAssertions;
-    let api: UndoLogSetupPublic;
-    let log: UndoLogPublic;
-  
-    beforeEach(async () => {
-      ({ logAssertions, logUtils, connection, api } = await setupBeforeEach());
-      log = await api.initializeSingle();
-    });
-  
-    afterEach(async () => {
-      await connection.close();
-    });
-
-    test('Causual usecase', () => {
-      //TODO
-    });
-});
