+++ src/undo-log.ts
@@ -1,5 +1,5 @@
 import { ChangeType, Channel } from "./undo-log-tables";
-import { ColumnValue } from "./utils";
+import { ColumnValue } from "./utilities";
 
 export class UndoLogError extends Error {
   constructor(message: string) {
