+++ src/utils/configuration-services.ts
@@ -1,0 +1,5 @@
+
+export interface ConfigurationServices {
+  getConfig(name: string): Promise<number>;
+  setConfig(name: string, value: number): Promise<void>;
+}
