+++ src/utils/private-services.ts
@@ -1,0 +1,7 @@
+import { Connection } from "../sqlite3";
+
+
+export interface PrivateServices {
+  connection: Connection;
+  prefix: string;
+}
